// 11-directiveSignals/src/app/signals/interfaces/user-request.interface.ts
//pegamos desde POSTMAN todos los datos de usario y con Paste JSON as Code le decimos TOP userSingleResponse// Generated by https://quicktype.io
// Generated by https://quicktype.io

//esta respuesta es lo que recibimos en Postman
export interface SingleUserResponse {
    data:        User; //cambiamos de Datum[] a User
    support:     Support;
}

export interface User { //cambiamos de Datum[] a User
    id:         number;
    email:      string;
    first_name: string;
    last_name:  string;
    avatar:     string;
}

export interface Support {
    url:  string;
    text: string;
}